Add authorization check

+++ controllers/feed.js
@@ -5,6 +5,7 @@ const { body } = require('express-validator/check');
 const { validationResult } = require('express-validator/check');
 
 const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
@@ -15,6 +16,7 @@ exports.getPosts = (req, res, next) => {
     .then(count => {
       totalItems = count;
       return Post.find()
+        .populate('creator', 'name')
         .sort({ createdAt: -1 })
         .skip((currentPage - 1) * POST_PER_PAGE)
         .limit(POST_PER_PAGE);
@@ -32,6 +34,7 @@ exports.getPosts = (req, res, next) => {
 exports.getSinglePost = (req, res, next) => {
   const postId = req.params.postId;
   Post.findById(postId)
+    .populate('creator', 'name')
     .then(post => {
       if (!post) {
         const error = new Error('Post not found.');
@@ -58,31 +61,41 @@ exports.createPost = (req, res, next) => {
   const imageUrl = req.file.path.replace('\\', '/');
   const title = req.body.title;
   const content = req.body.content;
+  let creator;
   const post = new Post({
     title: title,
     imageUrl: imageUrl,
     content: content,
-    creator: { name: 'Node.js' }
+    creator: req.userId
   });
   post
     .save()
     .then(result => {
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
+    .then(result => {
       res.status(201).json({
         message: 'Created a post successfully.',
-        post: result
+        post: post,
+        creator: { _id: creator._id, name: creator.name }
       });
     })
     .catch(next);
 };
 
 exports.updatePost = (req, res, next) => {
-  const postId = req.params.postId;
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
     const error = new Error('Validation failed. Entered data is incorrect.');
     error.statusCode = 422;
     throw error;
   }
+  const postId = req.params.postId;
   let imageUrl = req.body.image;
   if (req.file) {
     imageUrl = req.file.path.replace('\\', '/');
@@ -99,6 +112,11 @@ exports.updatePost = (req, res, next) => {
       if (!post) {
         const error = new Error('Post not found.');
         error.statusCode = 404;
+        throw error;
+      }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized.');
+        error.statusCode = 403;
         throw error;
       }
       if (imageUrl !== post.imageUrl) {
@@ -124,11 +142,22 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized.');
+        error.statusCode = 403;
+        throw error;
+      }
       clearImage(post.imageUrl);
       return Post.findByIdAndDelete(postId);
     })
     .then(result => {
-      console.log(result);
+      return User.findById(req.userId)
+    })
+    .then(user => {
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then(result => {
       res.status(200).json({ message: 'Post has been deleted.' });
     })
     .catch(next);
